refactoring
Комментирование полей/методов, исправление кодировки и пр.
