+++ GeometricLibrary/Shape.cs
@@ -1,7 +1,10 @@
-﻿namespace GeometricLibrary
+namespace GeometricLibrary
 {
     public abstract class Shape
     {
+        /// <summary>
+        /// Тип фигуры
+        /// </summary>
         public string TypeName { get; set; }
 
         public Shape(string typeName)
@@ -9,6 +12,9 @@
             TypeName = typeName;
         }
 
+        /// <summary>
+        /// Вычисление площади фигуры
+        /// </summary>
         public abstract double Square();
 
     }
