+++ README.md
@@ -8,6 +8,24 @@
 ●	Вычисление площади фигуры без знания типа фигуры в compile-time  
 ●	Проверку на то, является ли треугольник прямоугольным  
 
+## Задачи
+Поскольку нет тасктрекера, именование веток будет хаотичным, в зависимости от задачи 
+
+- Обязательные:  
+✅ Вычисление площади фигур (master)  
+✅ Возможность добавлять новые фигуры (master)  
+✅ Проверка на то, является ли треугольник прямоугольным (master)  
+⬜ Вычисление площади фигуры без знания типа фигуры в compile-time  
+✅ Базовые юнит-тесты (master) 
+
+
+- Опциональные:  
+✅ Добавить список задач в файл README.md (master) 
+✅ Проверка на то, удовлетворяет ли треугольник базовым правилам геометрии (master)  
+✅ Рефакторинг - комментирование полей/методов, исправление кодировки и пр. (refactoring)   
+⬜ Unit-tests - проверка треугольника на соответствие правилам геометрии  
+
+
 ## Вопрос №2
 В базе данных MS SQL Server есть продукты и категории. Одному продукту может соответствовать много категорий, в одной категории может быть много продуктов. Напишите SQL запрос для выбора всех пар «Имя продукта – Имя категории». Если у продукта нет категорий, то его имя все равно должно выводиться.
 По возможности — положите ответ рядом с кодом из первого вопроса.
