+++ UnitTest2/TriangleTest.cs
@@ -3,41 +3,32 @@ using Microsoft.VisualStudio.TestTools.UnitTesting;
 namespace GeometricLibrary.UnitTest
 {
     [TestClass]
-    public class CircleTests
-    {
-        [TestMethod]
-        public void Square_data()
-        {
-            var circle = new Circle("Круг", 13.3);
-            double expected = 555;
-
-            var result = Math.Round(circle.Square());
-
-            Assert.AreEqual(expected, result);
-        }
-    }
-
-    [TestClass]
-    public class TriangleTests
+    public class TriangleTest
     {
         [TestMethod]
         public void Triangle_data()
         {
+            // Arrange
             var triangle = new Triangle("Треугольник", 3, 4, 5);
             double expected = 6;
 
+            // Act
             var result = triangle.Square();
 
+            // Assert
             Assert.AreEqual(expected, result);
         }
 
         [TestMethod]
         public void isStraightTriangle_data()
         {
-            var triangle = new Triangle("Треугольник", 1, 2, 7);
+            // Arrange
+            var triangle = new Triangle("Треугольник", 2, 3, 4);
 
+            // Act
             var result = triangle.isStraightTriangle();
 
+            // Assert
             Assert.IsFalse(result);
         }
     }
